Rename rskip id and final value for minimumPayableGas
