+++ rskj-core/src/main/java/org/ethereum/config/blockchain/AbstractConfig.java
@@ -150,4 +150,7 @@ public abstract class AbstractConfig implements BlockchainConfig, BlockchainNetC
 
     @Override
     public boolean isRskip87() { return false; }
+    
+    @Override
+    public boolean isRskIp85() { return false; }
 }
