+++ rskj-core/src/main/java/org/ethereum/config/blockchain/regtest/RegTestGenesisConfig.java
@@ -70,5 +70,4 @@ public class RegTestGenesisConfig extends GenesisConfig {
     public boolean areBridgeTxsFree() {
         return true;
     }
-
 }
